Revert "[USERS] Endpoint pagination GetAllUsers #4"
