+++ package.json
@@ -17,7 +17,6 @@
     "node": ">=0.14"
   },
   "dependencies": {
-    "bcrypt": "^5.1.0",
     "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
     "debug": "~2.6.9",
@@ -25,7 +24,6 @@
     "express": "~4.16.1",
     "express-validator": "^6.13.0",
     "http-errors": "~1.6.3",
-    "jsonwebtoken": "^8.5.1",
     "morgan": "~1.9.1",
     "mysql2": "^2.3.3",
     "nodemon": "^2.0.15",
