--- routes/authRoutes.js
@@ -1,8 +1,0 @@
-const express = require('express')
-const { loginUser } = require('../controllers/auth/loginController');
-
-const router = express.Router()
-
-router.post('/auth/login', loginUser)
-
-module.exports = router
