+++ routes/index.js
@@ -1,9 +1,9 @@
 const express = require('express')
-const authRoutes = require('./authRoutes')
+const { get } = require('../controllers/index')
 
 const router = express.Router()
 
 // example of a route with index controller get function
-router.use("/", authRoutes);
+router.get('/', get)
 
 module.exports = router
